LayoutTests/editing/selection/move-selection-backwards-at-end-of-table-by-sentence-granularity-expected.txt

This tests moving the selection backwards at the end of a table by sentence granularity. The caret should move to the beginning of the table.
| "\n"
| <table>
|   "\n  "
|   <tbody>
|     <tr>
|       "\n    "
|       <th>
|         "<#selection-caret>First"
|       "\n    "
|       <th>
|         "Second"
|       " \n    "
|       <th>
|         id="target"
|         "Third"
|       "\n  "
|     "\n"
| "\n"
